--- app/src/main/java/woowacourse/movie/model/ui/MovieDisplayData.kt
@@ -1,3 +1,0 @@
-package woowacourse.movie.model.ui
-
-sealed class MovieDisplay(val viewType: Int)
