+++ app/src/main/java/woowacourse/movie/movieList/MovieItemListener.kt
@@ -1,7 +1,5 @@
 package woowacourse.movie.movieList
 
-import android.view.View
-
 fun interface MovieItemListener {
     fun onClick()
 }
