+++ app/src/main/java/woowacourse/movie/purchaseConfirmation/PurchaseConfirmationActivity.kt
@@ -5,10 +5,10 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.MenuItem
 import androidx.core.content.IntentCompat
-import woowacourse.movie.error.ErrorActivity
 import woowacourse.movie.R
 import woowacourse.movie.base.BindingActivity
 import woowacourse.movie.databinding.ActivityPurchaseConfirmationBinding
+import woowacourse.movie.error.ErrorActivity
 import woowacourse.movie.model.Cinema
 
 class PurchaseConfirmationActivity :
