+++ app/src/main/java/woowacourse/movie/seat/TheaterSeatPresenter.kt
@@ -77,5 +77,4 @@ class TheaterSeatPresenter(
             }
         view.showPrice(totalPrice)
     }
-
 }
