+++ app/src/test/java/woowacourse/movie/SeatPresenterTest.kt
@@ -113,6 +113,5 @@ class SeatPresenterTest {
         verify(exactly = 1) { view.setSeatBackground(seatId, colorSlot[1]) }
         colorSlot[0] shouldBe whiteColor
         colorSlot[1] shouldBe redColor
-
     }
 }
